Renaming cd_models to recourse methods as suggested in PR #19
