--- carla/cf_models/api/__init__.py
@@ -1,3 +1,0 @@
-# flake8: noqa
-
-from .cfmodel import CFModel
