--- carla/cf_models/catalog/dice/__init__.py
@@ -1,4 +1,0 @@
-# flake8: noqa
-
-# from .data_processing import prepare_dataset_for_dice
-from .model import Dice
