+++ carla/recourse_methods/api/__init__.py
@@ -1,0 +1,3 @@
+# flake8: noqa
+
+from .recourse_method import Recourse_Method
