+++ carla/recourse_methods/api/recourse_method.py
@@ -1,7 +1,7 @@
 from abc import ABC, abstractmethod
 
 
-class CFModel(ABC):
+class Recourse_Method(ABC):
     @abstractmethod
     def get_counterfactuals(self, data, mlmodel):
         pass
