+++ carla/recourse_methods/catalog/dice/__init__.py
@@ -1,0 +1,3 @@
+# flake8: noqa
+
+from .model import Dice
